+++ noodles-bam/examples/bam_flagstat.rs
@@ -5,7 +5,7 @@
 use std::{env, fmt, fs::File, io};
 
 use noodles_bam as bam;
-use noodles_sam::{self as sam, AlignmentRecord};
+use noodles_sam::{self as sam, alignment::Record, AlignmentRecord};
 
 #[derive(Debug, Default)]
 struct Counts {
@@ -27,7 +27,7 @@ struct Counts {
     mate_reference_sequence_id_mismatch_hq: u64,
 }
 
-fn count(counts: &mut Counts, record: &bam::Record) {
+fn count(counts: &mut Counts, record: &Record) {
     let flags = record.flags();
 
     counts.read += 1;