+++ noodles-bam/src/lib.rs
@@ -58,7 +58,7 @@ pub mod reader;
 pub mod record;
 pub mod writer;
 
-pub use self::{reader::Reader, record::Record, writer::Writer};
+pub use self::{reader::Reader, writer::Writer};
 
 #[cfg(feature = "async")]
 pub use self::r#async::{Reader as AsyncReader, Writer as AsyncWriter};